+++ global/factory/types.ts
@@ -8,10 +8,15 @@ export type AttributeOperator = 'prefixed' | 'suffixed' | 'contains';
 /**
  * Global params.
  */
-export interface GlobalAttributeParams {
+export interface GlobalAttributeParams<AttributeKeys extends Record<string, string>> {
   /**
    * Defines if the `<script>` should prevent automatically loading the library.
    * Useful for cases where a JS developer whants to programatically init the library.
    */
   preventsLoad: boolean;
+
+  /**
+   * The parsed custom attributes from the <script> tag.
+   */
+  attributes: { [Key in keyof AttributeKeys]: string | null };
 }