+++ packages/a11y/src/index.ts
@@ -1,7 +1,23 @@
-import { observeAriaControls } from './aria/aria-controls';
-import { emitClickEvents } from './keyboard/keyboard-clicks';
+import { assessScript, initAttributes } from '@global/factory';
+import { A11Y_ATTRIBUTE } from 'global/constants/attributes';
 
-document.addEventListener('DOMContentLoaded', () => {
-  emitClickEvents();
-  observeAriaControls();
-});
+import { version } from '../package.json';
+import { init } from './init';
+
+/**
+ * Init
+ */
+initAttributes();
+
+window.fsAttributes[A11Y_ATTRIBUTE] ||= {};
+
+const { preventsLoad } = assessScript();
+const attribute = window.fsAttributes[A11Y_ATTRIBUTE];
+
+attribute.version = version;
+
+if (preventsLoad) attribute.init = init;
+else {
+  window.Webflow ||= [];
+  window.Webflow.push(init);
+}