+++ packages/socialshare/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @finsweet/attributes-socialshare
 
+## 1.0.3
+
+### Patch Changes
+
+- fa1447f8: Added case Insensitive and fix leak that led to other elements with the same prefix
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [7542dcdb]
+  - @global/constants@1.0.2
+  - @global/factory@1.1.3
+
+## 1.0.1
+
+### Patch Changes
+
+- 7d7fb30: Redefined required fields, content is no longer mandatory. Code improvements.
+- Updated dependencies [bdd1a78]
+  - @global/constants@1.0.1
+  - @global/factory@1.1.2
+
 ## 1.0.0
 
 ### Major Changes